move teamGetByItemID into app/api/groups
