+++ app/api/currentuser/current_user.go
@@ -43,8 +43,6 @@ func (srv *Service) SetRoutes(router chi.Router) {
 
 	router.Get("/current-user/full-dump", service.AppHandler(srv.getFullDump).ServeHTTP)
 	router.Get("/current-user/dump", service.AppHandler(srv.getDump).ServeHTTP)
-
-	router.Get("/current-user/teams/by-item/{item_id}", service.AppHandler(srv.getTeamByItem).ServeHTTP)
 }
 
 type userGroupRelationAction string
