+++ app/api/groups/get_current_user_team_by_item.feature
