+++ app/api/groups/get_current_user_team_by_item.go
@@ -1,4 +1,4 @@
-package currentuser
+package groups
 
 import (
 	"errors"
@@ -46,7 +46,7 @@ import (
 //     "$ref": "#/responses/notFoundResponse"
 //   "500":
 //     "$ref": "#/responses/internalErrorResponse"
-func (srv *Service) getTeamByItem(w http.ResponseWriter, r *http.Request) service.APIError {
+func (srv *Service) getCurrentUserTeamByItem(w http.ResponseWriter, r *http.Request) service.APIError {
 	itemID, err := service.ResolveURLQueryPathInt64Field(r, "item_id")
 	if err != nil {
 		return service.ErrInvalidRequest(err)
